fix optaplanner

+++ ospp-application/src/main/java/ch/fhnw/ip6/ospp/service/StudentService.java
@@ -1,6 +1,8 @@
 package ch.fhnw.ip6.ospp.service;
 
+import ch.fhnw.ip6.ospp.model.Presentation;
 import ch.fhnw.ip6.ospp.model.Student;
+import ch.fhnw.ip6.ospp.persistence.PresentationRepository;
 import ch.fhnw.ip6.ospp.persistence.StudentRepository;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.log4j.Log4j2;
@@ -15,12 +17,22 @@ import java.util.Optional;
 public class StudentService {
 
     private final StudentRepository studentRepository;
+    private final PresentationRepository presentationRepository;
 
     public Student save(Student student) {
         return studentRepository.save(student);
     }
 
     public void delete(Long id) {
+        Optional<Student> student = studentRepository.findById(id);
+
+        if (!student.isPresent()){
+            throw new FachlicheException("Schüler existiert nicht.");
+        }
+        List<Presentation> presentations = presentationRepository.findByStudentOneOrStudentTwo(student.get(),student.get());
+        if(presentations.size() > 0){
+            throw new FachlicheException("Dem Schüler sind noch Präsentationen zugewiesen");
+        }
         studentRepository.deleteById(id);
     }
